# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DramaShort API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: string;
  displayName: string;
  avatarUrl: string;
  coins: number;
  favoritesCount: number;
  watchedCount: number;
  memberSince: string;
}

export interface CoinPack {
  id: string;
  coins: number;
  priceUsd: number;
  bonus: number;
  label: string;
}

export interface Drama {
  id: string;
  title: string;
  description: string;
  category: string;
  genre: string[];
  releaseYear: number;
  coverUrl: string;
  bannerUrl: string;
  rating: number;
  viewsCount: number;
  episodesCount: number;
  durationLabel: string;
  isNew: boolean;
  isHot: boolean;
}

export interface Episode {
  id: string;
  dramaId: string;
  episodeNumber: number;
  title: string;
  durationSeconds: number;
  thumbnailUrl: string;
  videoUrl: string;
  isLocked: boolean;
  unlockCost: number;
}

export type DramaDetail = Drama & {
  episodes: Episode[];
  isFavorite: boolean;
};

export interface EpisodePlayback {
  episode: Episode;
  drama: Drama;
  prevEpisodeId?: string | null;
  nextEpisodeId?: string | null;
  progressSeconds: number;
}

export interface UnlockResult {
  success: boolean;
  episode: Episode;
  coinsRemaining: number;
}

export interface Genre {
  name: string;
  count: number;
}

export interface ContinueWatchingItem {
  drama: Drama;
  episode: Episode;
  progressSeconds: number;
  progressPercent: number;
  updatedAt: string;
}

export type PurchaseCoinsBody = {
  packId: string;
};

export type ListDramasParams = {
  search?: string;
  genre?: string;
  sort?: ListDramasSort;
};

export type ListDramasSort =
  (typeof ListDramasSort)[keyof typeof ListDramasSort];

export const ListDramasSort = {
  popular: "popular",
  newest: "newest",
  trending: "trending",
} as const;

export type AddFavoriteBody = {
  dramaId: string;
};

export type SaveProgressBody = {
  episodeId: string;
  progressSeconds: number;
};