# lib/api-zod/src/generated/types/coinPack.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DramaShort API
 * OpenAPI spec version: 0.1.0
 */

export interface CoinPack {
  id: string;
  coins: number;
  priceUsd: number;
  bonus: number;
  label: string;
}