# lib/api-zod/src/generated/types/continueWatchingItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DramaShort API
 * OpenAPI spec version: 0.1.0
 */
import type { Drama } from "./drama";
import type { Episode } from "./episode";

export interface ContinueWatchingItem {
  drama: Drama;
  episode: Episode;
  progressSeconds: number;
  progressPercent: number;
  updatedAt: string;
}