# lib/api-zod/src/generated/types/dramaDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DramaShort API
 * OpenAPI spec version: 0.1.0
 */
import type { Drama } from "./drama";
import type { Episode } from "./episode";

export type DramaDetail = Drama & {
  episodes: Episode[];
  isFavorite: boolean;
};