# lib/api-zod/src/generated/types/episode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DramaShort API
 * OpenAPI spec version: 0.1.0
 */

export interface Episode {
  id: string;
  dramaId: string;
  episodeNumber: number;
  title: string;
  durationSeconds: number;
  thumbnailUrl: string;
  videoUrl: string;
  isLocked: boolean;
  unlockCost: number;
}