# lib/api-zod/src/generated/types/unlockResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DramaShort API
 * OpenAPI spec version: 0.1.0
 */
import type { Episode } from "./episode";

export interface UnlockResult {
  success: boolean;
  episode: Episode;
  coinsRemaining: number;
}