# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DramaShort API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: string;
  displayName: string;
  avatarUrl: string;
  coins: number;
  favoritesCount: number;
  watchedCount: number;
  memberSince: string;
}